Move all enums files to the enums directory
