+++ implementation/src/backend/user-management/user-management/src/main/java/com/matjar/user/enums/ActivationCodeStatus.java
@@ -1,0 +1,6 @@
+package com.matjar.user.enums;
+
+public enum ActivationCodeStatus {
+    UNTOUCHED,
+    USED
+}
