+++ implementation/src/backend/user-management/user-management/src/main/java/com/matjar/user/enums/LoggingStatus.java
@@ -1,0 +1,8 @@
+package com.matjar.user.enums;
+
+public enum LoggingStatus {
+    STARTED,
+    COMPLETED,
+    SUCCESS,
+    FAILED
+}
