+++ implementation/src/backend/user-management/user-management/src/main/java/com/matjar/user/enums/LoginStatus.java
@@ -1,4 +1,4 @@
-package com.matjar.user.model;
+package com.matjar.user.enums;
 
 public enum LoginStatus {
     SUCCESS,
